query-builder: reorganize code (part 2)
