+++ packages/legend-application-query/src/stores/fetch-structure/projection/aggregation/operators/QueryBuilderAggregateOperatorHelper.ts
@@ -24,11 +24,9 @@ import {
   TYPICAL_MULTIPLICITY_TYPE,
 } from '@finos/legend-graph';
 import { guaranteeType, assertTrue } from '@finos/legend-shared';
-import {
-  type QueryBuilderAggregateOperator,
-  QueryBuilderAggregateColumnState,
-} from '../QueryBuilderAggregationState.js';
+import { QueryBuilderAggregateColumnState } from '../QueryBuilderAggregationState.js';
 import type { QueryBuilderProjectionColumnState } from '../../QueryBuilderProjectionState.js';
+import type { QueryBuilderAggregateOperator } from '../QueryBuilderAggregateOperator.js';
 
 export const buildAggregateExpression = (
   operatorFunctionFullPath: string,
