+++ packages/legend-application-query/src/stores/filter/operators/QueryBuilderFilterOperator_StartWith.ts
